Refactor KafkaApis a bit and add ControllerApis
