+++ core/src/main/scala/kafka/server/KafkaConfig.scala
@@ -1041,7 +1041,7 @@ object KafkaConfig {
       .define(ControllerConnectProp, LIST, null, HIGH, ControllerConnectDoc)
       .define(RegistrationHeartbeatIntervalMsProp, INT, Defaults.RegistrationHeartbeatIntervalMs, MEDIUM, RegistrationHeartbeatIntervalMsDoc)
       .define(RegistrationLeaseTimeoutMsProp, INT, Defaults.RegistrationLeaseTimeoutMs, MEDIUM, RegistrationLeaseTimeoutMsDoc)
-      .define(MetadataLogDirProp, null, HIGH, MetadataLogDirDoc)
+      .define(MetadataLogDirProp, STRING, HIGH, MetadataLogDirDoc)
 
       /************* Authorizer Configuration ***********/
       .define(AuthorizerClassNameProp, STRING, Defaults.AuthorizerClassName, LOW, AuthorizerClassNameDoc)
