+++ core/src/main/scala/kafka/server/KafkaProcess.scala
@@ -1,0 +1,5 @@
+package kafka.server
+
+class KafkaProcess {
+
+}
